move get obs missing
